requirements.txt
fastapi
sqlalchemy
pynput
pillow
uvicorn
pystray